+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Zappa Changelog
 
+## 0.41.3
+* Various Python3 fixes
+* Remove some dead code
+* More binary package fixes thanks to and @nikbora and @bxm156
+* Improved async tasks thanks to @schuyler1d
+* Various small changes
+
+## 0.41.2
+* Support for new `lambda-packages` format (Python3 support)
+* Fix `setup.py` on Windows Python3
+* Fix #818 - python3 import for LE
+* Support AWS-specific environment variables (with KMS)
+
+## 0.41.1
+* Add `template` command
+* Add `--json` in more places
+* Add `--output` to package
+* Support for manylinux wheels Python 3.6 package downloading #803
+from nikbora
+* Fix PyEnv exit code #799
+
+## 0.41.0
+* Add Python3 Support! #793, #6
+* Deprecate `lets_encrypt_expression`
+* Refactor a bunch of stuff to work with Python3 package restrictions >:[
+* #776 fix for libmysqlclient.so.18 error when using `slim_handler`
+* add profile and region detection to init - thanks @pdpol
+* #774 Wsgi environment improvements (Fix untrustworthy remote_addr)
+* Only create `__init__.py` file if there are python files or sub dirs in the folder
+* Update docs to reflect lambda name prepended to role_name
+* Guard log responses (thanks @scoates)
+
+## 0.40.0
+* Add Async Task Support! Lots of tickets and PRs related, including #61, #603, #694 and #732.
+* More info here: https://blog.zappa.io/posts/zappa-introduces-seamless-asynchronous-task-execution
+* Fix Django non-WSGI function initialization, #748
+* Add support for AWS Lambda Dead Letter Queue, #740
+* Fix API Gateway test button (the bolt button), #735
+* Switch to using per-lambda-name (project-stage) rather than a single default LambdaExecutionRole
+
 ## 0.39.1
 * Fix broken Let's Encrypt trying to use new ACM ARNs
 * Add `apigateway_description` setting, fixes #722